+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Updated Node.js support to `^12.22.0 || ^14.17.0 || >= 16.0.0`.
 - Updated dev dependencies, some of which require newer Node.js versions than previously supported.
+- Removed `./package` from the package `exports` field; the full `package.json` filename must be used in a `require` path.
 
 ### Patch